chore: add wrapper for telemetry
